feat(flake.nix): add nixpkgs-master input and use it for spotify-player
This commit introduces the `nixpkgs-master` input to the `flake.nix` file.
It also updates the `home/mrzelee/packages/media/default.nix` to use `master.spotify-player` instead of `unstable.spotify-player`.
This change is made to leverage the latest version of `spotify-player` available in the master branch of nixpkgs, which may include newer features or bug fixes.
The `modules/nixpkgs-overlays.nix` file is updated to correctly import and configure the `nixpkgs-master` overlay.