feat(network): add tailscale service and configure DNS and hosts
This commit introduces the Tailscale service to the system configuration.
It also configures the system to use Google DNS and Cloudflare DNS as
nameservers, and adds an entry for 'azurite' to the hosts file.
The Tailscale service is enabled with client routing features and
specific nameservers and search domains are configured for Tailscale
connectivity.