feat(hyprland): enable hyprlock and add hyprpaper to system packages
This commit enables the hyprlock program for Hyprland and adds hyprpaper to the system packages. This provides users with a locking mechanism and wallpaper management for their Hyprland environment.

Additionally, this commit updates the GTK configuration for the jmoura user. It sets the preference for dark themes, disables cursor blinking, and configures the recent files limit. It also adds several common directories to the GTK bookmarks for easier access.

Finally, the swaylock package has been removed from the wayland.nix configuration as it is no longer needed with the introduction of hyprlock.