chore(deps): update nix flake inputs and package versions
This commit updates several Nix flake inputs to newer versions, including nixpkgs, nix-darwin, home-manager, and nixos-apple-silicon. It also updates package versions for Julia and Zulu.

Additionally, it refactors the boot configuration for the macbook-nixos to use a labeled swap device instead of a swap file, which is a more robust approach for hibernation. The experimental GPU driver for Asahi Linux has been removed as it's no longer experimental. The logind settings have been updated to use the new NixOS module syntax. Finally, the hardware configuration has been updated to reflect the LUKS encryption setup for the root partition.